api/requirements.txt
coverage==4.5.4
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
psycopg2-binary==2.8.3
gunicorn==19.9.0
flask-bcrypt==0.7.1
flask-cors==3.0.8
flask-marshmallow==0.10.1
flask-migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Testing==0.7.1
marshmallow-sqlalchemy==0.17.0
pyjwt==1.7.1
requests==2.20.1
SQLAlchemy-Utils==0.34.1
